Release new versions of the gem upon tag pushes
When a new tag is pushed to GitHub and verified by Travis CI, deploy
that gem to RubyGems.org so that people can begin using it immediately.
This allows for any member of the team to create a new gem release
simply by having the access to push a tag to the GitHub repository.